chore(test): render snapshot for build and dev independently (#6584)
- `build_artifacts_snapshot.rs`​clearly shows how it renders the snapshot
- I didn't put much attention on `dev_artifacts_snapshot.rs `​, becuase it needs be refactored to satisfy more scenarios (#6567) later.
- Another reason for splitting the logic is that `dev`​ test gonna more complex logic than normal build. I feel the disadvantages it brings are beyond advatanges for unifying them.
