+++ crates/rolldown_testing/src/types/mod.rs
@@ -1,7 +1,11 @@
-mod artifacts_snapshot;
+mod build_artifacts_snapshot;
 mod build_round_output;
+mod dev_artifacts_snapshot;
+mod dev_round_output;
 mod snapshot_section;
 
-pub use artifacts_snapshot::ArtifactsSnapshot;
+pub use build_artifacts_snapshot::BuildArtifactsSnapshot;
 pub use build_round_output::BuildRoundOutput;
+pub use dev_artifacts_snapshot::DevArtifactsSnapshot;
+pub use dev_round_output::DevRoundOutput;
 pub use snapshot_section::SnapshotSection;
